Assignments/Assignment1/requirements.txt
matplotlib==3.8.4
numpy==1.21.5
pandas==2.2.2
Pillow==10.3.0
scikit_learn==1.4.2
tensorflow==2.16.1
